Updating files from the main branch

+++ Fibonacci.java
@@ -7,7 +7,7 @@ public class Fibonacci {
         BigInteger num1 = new BigInteger("0");
         BigInteger num2 = new BigInteger("1");
 
-        for(long loop = 0; loop < n; loop++) {
+        for(int loop = 0; loop < n; loop++) {
             fibonacci = num1.add(num2);
             num1 = num2;
             num2 = fibonacci;
